Convert all dxtbx tests to pytest
All tests in dxtbx are now pytest compliant. Testing with
libtbx.run_tests_parallel will still be supported for a while.

Tests were split up where appropriate, which is why the number of tests
increases from 70 to 275.

To run dxtbx tests run
  $ pytest
while being inside the dxtbx directory.

For more information about pytest and how to write pytests please see
  https://github.com/dials/dials/wiki/pytest
