+++ dxtbx/tests/model/test_detector.py
@@ -1,4 +1,4 @@
-from __future__ import absolute_import, division
+from __future__ import absolute_import, division, print_function
 
 from dxtbx.model import Detector, Panel
 from libtbx.test_utils import approx_equal
@@ -52,7 +52,6 @@ def tst_is_coord_valid(detector):
   assert(detector[0].is_coord_valid((256, 513)) == False)
 
 def tst_pixel_to_millimeter_to_pixel(detector):
-
   from scitbx import matrix
   from random import random
   eps = 1e-7
@@ -129,7 +128,6 @@ def tst_detectors_are_different(detA, detB):
   assert(detA != detB)
 
 def tst_resolution(detector):
-
   from dxtbx.model import Beam
   beam = Beam(direction=(0,0,1), wavelength=1.0)
   d_min1 = detector.get_max_resolution(beam.get_s0())
@@ -173,7 +171,7 @@ def tst_panel_mask():
   assert m2[30,30] == False
   assert m2[40,40] == False
 
-def tst_detector():
+def test_detector():
   from dxtbx.model import ParallaxCorrectedPxMmStrategy
 
   def create_detector(offset = 0):
@@ -238,9 +236,3 @@ def tst_detector():
       ParallaxCorrectedPxMmStrategy(mu, t0)))
 
   tst_parallax_correction(detector)
-
-def run():
-  tst_detector()
-
-if __name__ == '__main__':
-  run()
