+++ dxtbx/tests/model/test_ray_intersection.py
@@ -1,7 +1,8 @@
-from __future__ import absolute_import, division
+from __future__ import absolute_import, division, print_function
+
+import pytest
 
 def tst_intersection_at_origin(intersection, wavelength, origin):
-
   eps = 1e-7
 
   # Set the beam vector to directed towards the detector origin
@@ -11,11 +12,10 @@ def tst_intersection_at_origin(intersection, wavelength, origin):
   x, y = intersection(s)
 
   # Check that they are equal to the detector origin
-  assert(abs(x) < eps)
-  assert(abs(y) < eps)
+  assert abs(x) < eps
+  assert abs(y) < eps
 
 def tst_intersection_at_corners(intersection, panel, wavelength):
-
   from scitbx import matrix
   eps = 1e-7
 
@@ -39,29 +39,22 @@ def tst_intersection_at_corners(intersection, panel, wavelength):
   x4, y4 = intersection(s4)
 
   # Check that they are equal to the detector origin
-  assert(abs(x1 - 0) < eps and abs(y1 - 0) < eps)
-  assert(abs(x2 - image_size[0]) < eps and abs(y2 - 0) < eps)
-  assert(abs(x3 - 0) < eps and abs(y3 - image_size[1]) < eps)
-  assert(abs(x4 - image_size[0]) < eps and abs(y4 - image_size[1]) < eps)
+  assert abs(x1 - 0) < eps and abs(y1 - 0) < eps
+  assert abs(x2 - image_size[0]) < eps and abs(y2 - 0) < eps
+  assert abs(x3 - 0) < eps and abs(y3 - image_size[1]) < eps
+  assert abs(x4 - image_size[0]) < eps and abs(y4 - image_size[1]) < eps
 
 def tst_intersection_away_from_panel(intersection, panel, wavelength):
-
   from scitbx import matrix
 
   # Set the beam vector to directed away from the detector origin
   s = - (1.0 / wavelength) * matrix.col(panel.get_origin())
 
   # Check that they are equal to None
-  try:
+  with pytest.raises(RuntimeError):
     intersection(s)
-  except(RuntimeError):
-    # Test Passed
-    return
 
-  # Test Failed
-  assert(False)
-
-def tst_beam_plane_intersection():
+def test_beam_plane_intersection():
   from dxtbx.model import Panel
   from scitbx import matrix
 
@@ -87,7 +80,7 @@ def tst_beam_plane_intersection():
                 image_size, (0, 0), 0.0, "")
 
   # Create the intersection object
-  intersection = lambda x: panel.get_ray_intersection(x)
+  intersection = panel.get_ray_intersection
 
   # Perform a load of tests
   tst_intersection_at_origin(intersection, wavelength, origin)
@@ -95,16 +88,15 @@ def tst_beam_plane_intersection():
   tst_intersection_away_from_panel(intersection, panel, wavelength)
 
 def tst_transform_at_origin(transform, panel):
-
   eps = 1e-7
 
   # Get lab coordinate at point
   x, y, z = transform((0, 0))
 
   # Check coordinate is at origin
-  assert(abs(x - panel.get_origin()[0]) < eps)
-  assert(abs(y - panel.get_origin()[1]) < eps)
-  assert(abs(z - panel.get_origin()[2]) < eps)
+  assert abs(x - panel.get_origin()[0]) < eps
+  assert abs(y - panel.get_origin()[1]) < eps
+  assert abs(z - panel.get_origin()[2]) < eps
 
 def tst_transform_at_corners(transform, panel):
   from scitbx import matrix
@@ -129,12 +121,12 @@ def tst_transform_at_corners(transform, panel):
   xyz44 = matrix.col(transform((image_size[0], image_size[1])))
 
   # Check that they are equal to the detector origin
-  assert(abs(xyz11 - xyz1) < eps)
-  assert(abs(xyz22 - xyz2) < eps)
-  assert(abs(xyz33 - xyz3) < eps)
-  assert(abs(xyz44 - xyz4) < eps)
+  assert abs(xyz11 - xyz1) < eps
+  assert abs(xyz22 - xyz2) < eps
+  assert abs(xyz33 - xyz3) < eps
+  assert abs(xyz44 - xyz4) < eps
 
-def tst_plane_to_lab_transform():
+def test_plane_to_lab_transform():
   from dxtbx.model import Panel
   from scitbx import matrix
 
@@ -160,7 +152,7 @@ def tst_plane_to_lab_transform():
                image_size, (0, 0), 0.0, "")
 
   # Create the intersection object
-  transform = lambda x: panel.get_lab_coord(x)
+  transform = panel.get_lab_coord
 
   # Perform a load of tests
   tst_transform_at_origin(transform, panel)
@@ -174,7 +166,8 @@ def tst_fwd_rev_random(intersection, transform, panel):
 
   # Set the shift parameters
   image_size = panel.get_image_size_mm()
-  random_xy = lambda: (random() * image_size[0], random() * image_size[1])
+  def random_xy():
+    return (random() * image_size[0], random() * image_size[1])
 
   # Loop a number of times
   num = 1000
@@ -190,9 +183,9 @@ def tst_fwd_rev_random(intersection, transform, panel):
     xy_2 = intersection(s)
 
     # Check the vectors are almost equal
-    assert(abs(xy - matrix.col(xy_2)) < eps)
+    assert abs(xy - matrix.col(xy_2)) < eps
 
-def tst_forward_and_reverse_transform():
+def test_forward_and_reverse_transform():
   from dxtbx.model import Panel
   from scitbx import matrix
 
@@ -218,16 +211,8 @@ def tst_forward_and_reverse_transform():
                image_size, (0, 0), 0.0, "")
 
   # Create the intersection object and transform object
-  intersection = lambda x: panel.get_ray_intersection(x)
-  transform = lambda x: panel.get_lab_coord(x)
+  intersection = panel.get_ray_intersection
+  transform = panel.get_lab_coord
 
   # Do a test
   tst_fwd_rev_random(intersection, transform, panel)
-
-def run():
-  tst_beam_plane_intersection()
-  tst_plane_to_lab_transform()
-  tst_forward_and_reverse_transform()
-
-if __name__ == '__main__':
-  run()
