+++ dxtbx/tests/model/test_scan_helpers.py
@@ -1,14 +1,15 @@
-from __future__ import absolute_import, division
+from __future__ import absolute_import, division, print_function
+
 from dxtbx.model import is_angle_in_range
 from dxtbx.model import get_range_of_mod2pi_angles
 from dxtbx.model import get_mod2pi_angles_in_range
 
-def tst_is_angle_in_range():
+# Run tests for the scan_helpers.h module.
+
+def test_is_angle_in_range():
   """Test that for a range of angles and angular ranges, the
   is_angle_in_range function correctly calculates if the angle
-  is in the range.
-
-  """
+  is in the range."""
   from random import random
 
   # Some helper lambda functions
@@ -30,11 +31,11 @@ def tst_is_angle_in_range():
       #   angle in range A -> B = True
       #   angle in range B -> 360 = False
       for angle in range(0, angular_range[0]):
-        assert(is_angle_in_range(angular_range, angle, True) == False)
+        assert is_angle_in_range(angular_range, angle, True) is False
       for angle in range(angular_range[0], angular_range[1]+1):
-        assert(is_angle_in_range(angular_range, angle, True) == True)
+        assert is_angle_in_range(angular_range, angle, True) is True
       for angle in range(angular_range[1]+1, 360):
-        assert(is_angle_in_range(angular_range, angle, True) == False)
+        assert is_angle_in_range(angular_range, angle, True) is False
     else:
 
       # Check that the following are true
@@ -42,62 +43,49 @@ def tst_is_angle_in_range():
       #   angle in range B -> A = False
       #   angle in range A -> 360 = True
       for angle in range(0, angular_range[1]+1):
-        assert(is_angle_in_range(angular_range, angle, True) == True)
+        assert is_angle_in_range(angular_range, angle, True) is True
       for angle in range(angular_range[1]+1, angular_range[0]):
-        assert(is_angle_in_range(angular_range, angle, True) == False)
+        assert is_angle_in_range(angular_range, angle, True) is False
       for angle in range(angular_range[0], 360):
-        assert(is_angle_in_range(angular_range, angle, True) == True)
+        assert is_angle_in_range(angular_range, angle, True) is True
 
   # Create a range over 360 and make sure all angles are valid
   angular_range = (-10, 370)
   for angle in range(0, 360):
-    assert(is_angle_in_range(angular_range, angle, True) == True)
+    assert is_angle_in_range(angular_range, angle, True) is True
 
-def tst_get_range_of_mod2pi_angles():
-  """Get the range of equivalent within a given angular range.
-
-  """
+def test_get_range_of_mod2pi_angles():
+  """Get the range of equivalent within a given angular range."""
   # In a 360 deg range, have only 1 angle
   a0, a1 = get_range_of_mod2pi_angles((0, 360), 180, deg=True)
-  assert(a0 == 180 and a1 == 180)
+  assert a0 == 180 and a1 == 180
 
   # If no angles within range, a0 > a1
   a0, a1 = get_range_of_mod2pi_angles((181, 360), 180, deg=True)
-  assert(a0 > a1)
+  assert a0 > a1
 
   # With 720 deg range, have 2 angles
   a0, a1 = get_range_of_mod2pi_angles((0, 720), 180, deg=True)
-  assert(a0 == 180 and a1 == 540)
+  assert a0 == 180 and a1 == 540
 
   # With 1080 deg range, have 3 angles
   a0, a1 = get_range_of_mod2pi_angles((-360, 720), 180, deg=True)
-  assert(a0 == -180 and a1 == 540)
+  assert a0 == -180 and a1 == 540
 
-def tst_get_mod2pi_angles_in_range():
-  """Get the list of equivalent within a given angular range.
-
-  """
+def test_get_mod2pi_angles_in_range():
+  """Get the list of equivalent within a given angular range."""
   # In a 360 deg range, have only 1 angle
   a = get_mod2pi_angles_in_range((0, 360), 180, deg=True)
-  assert(len(a) == 1 and a[0] == 180)
+  assert len(a) == 1 and a[0] == 180
 
   # If no angles within range have no angles
   a = get_mod2pi_angles_in_range((181, 360), 180, deg=True)
-  assert(len(a) == 0)
+  assert len(a) == 0
 
   # With 720 deg range, have 2 angles
   a = get_mod2pi_angles_in_range((0, 720), 180, deg=True)
-  assert(len(a) == 2 and a[0] == 180 and a[1] == 540)
+  assert len(a) == 2 and a[0] == 180 and a[1] == 540
 
   # With 1080 deg range, have 3 angles
   a = get_mod2pi_angles_in_range((-360, 720), 180, deg=True)
-  assert(len(a) == 3 and a[0] == -180 and a[1] == 180 and a[2] == 540)
-
-def run():
-  """Run tests for the scan_helpers.h module."""
-  tst_is_angle_in_range()
-  tst_get_range_of_mod2pi_angles()
-  tst_get_mod2pi_angles_in_range()
-
-if __name__ == '__main__':
-  run()
+  assert len(a) == 3 and a[0] == -180 and a[1] == 180 and a[2] == 540
