+++ dxtbx/tests/test_beamline_definitions.py
@@ -1,17 +1,7 @@
-#!/usr/bin/env python
-#
-#   Copyright (C) 2016 Diamond Light Source, Markus Gerstel
-#
-#   This code is distributed under the BSD license, a copy of which is
-#   included in the root directory of this package.
-#
-# Tests for the beamline definition database
-
-from __future__ import absolute_import, division
-
-import dxtbx.data.beamline_defs as dxbd
+from __future__ import absolute_import, division, print_function
 
 def test_lookup_unknown_detector():
+  import dxtbx.data.beamline_defs as dxbd
   n = dxbd.get_beamline_definition('This is a detector serial number that does not exist')
   assert 'Dummy CIF' in str(n), n
   assert str(n.CIF_block()) == ''
@@ -20,16 +10,13 @@ def test_lookup_unknown_detector():
   assert n.mmCIF_block().__module__ == 'iotbx.cif.model'
 
 def test_lookup_known_detector():
+  import dxtbx.data.beamline_defs as dxbd
   n = dxbd.get_beamline_definition('PILATUS 2M, S/N 24-0107 Diamond')
   assert 'Dummy' not in str(n), n
   assert str(n) != ''
   cif = n.CIF_block()
   assert cif.__module__ == 'iotbx.cif.model'
-  assert '_diffrn_radiation_type' in str(cif), cif
+  assert '_diffrn_radiation_type' in str(cif)
   cif = n.mmCIF_block()
   assert cif.__module__ == 'iotbx.cif.model'
-  assert '_diffrn_radiation.type' in str(cif), cif
-
-if __name__ == '__main__':
-  test_lookup_unknown_detector()
-  test_lookup_known_detector()
+  assert '_diffrn_radiation.type' in str(cif)
