+++ dxtbx/tests/test_dials_226.py
@@ -1,16 +1,13 @@
-from __future__ import absolute_import, division
+from __future__ import absolute_import, division, print_function
 
 import os
+import pytest
 
-import libtbx.load_env
 from dxtbx.datablock import DataBlockFactory
 
-def run():
+@pytest.mark.skip("disabled by GW 2016/OCT/17 to enable nightly builds to pass")
+def test(dials_regression):
   # https://github.com/dials/dials/issues/226
-  if not libtbx.env.has_module("dials_regression"):
-    print "Skipping tst_dials_226.py: dials_regression not present"
-    return
-  dials_regression = libtbx.env.dist_path('dials_regression')
   h5_path = os.path.join(
     dials_regression, 'image_examples/SACLA_MPCCD_Cheetah/run266702-0-subset.h5')
   datablock = DataBlockFactory.from_filenames([h5_path])[0]
@@ -20,8 +17,3 @@ def run():
   sliced_imageset = imageset[2:]
   # this line fails
   path = sliced_imageset.get_path(sliced_imageset.indices()[0])
-
-
-if __name__ == '__main__':
-  #run() # disabled 2016/OCT/17 to enable nightly builds to pass
-  pass # I guess
