--- dxtbx/tests/tst_dxtbx.py
@@ -1,127 +1,0 @@
-from __future__ import absolute_import, division
-
-def tst_dxtbx():
-  import libtbx.load_env
-  if not libtbx.env.has_module("dials"):
-    print "Skipping test: dials not present"
-    return
-  try:
-    dials_regression = libtbx.env.dist_path('dials_regression')
-  except KeyError:
-    print 'FAIL: dials_regression not configured'
-    return
-
-  import os
-  #from boost.python import streambuf
-  #from dxtbx import read_uint16
-  from dxtbx.format.Registry import Registry
-
-  from dials_regression.image_examples.get_all_working_images import \
-      get_all_working_images
-
-  for directory, image in get_all_working_images():
-    file_path = os.path.join(dials_regression, 'image_examples',
-                             directory, image)
-    format = Registry.find(file_path)
-    i = format(file_path)
-    det = i.get_detector()
-    if det is not None:
-      size = det[0].get_image_size()
-    b = i.get_beam()
-    g = i.get_goniometer()
-    s = i.get_scan()
-    try:
-      d = i.get_raw_data()
-    except IOError:
-      pass
-
-
-def tst_dxtbx_compressed():
-  import libtbx.load_env
-  try:
-    dials_regression = libtbx.env.dist_path('dials_regression')
-  except KeyError:
-    print 'FAIL: dials_regression not configured'
-    return
-
-  import os
-  from dxtbx.format.Registry import Registry
-
-  from dials_regression.image_examples.get_all_working_images import \
-      get_all_working_images
-
-  # test that reading gz or bz2 compressed files works: it doesn't!
-
-  from libtbx import smart_open
-  from libtbx.test_utils import open_tmp_directory
-  import shutil
-
-  tmp_dir = open_tmp_directory()
-  print tmp_dir
-
-  for directory, image in get_all_working_images():
-    file_path = os.path.join(dials_regression, 'image_examples',
-                             directory, image)
-    for ext in ('.gz', '.bz2')[:]:
-      compressed_path = os.path.join(tmp_dir, os.path.basename(file_path)) + ext
-      with open(file_path, 'rb') as f_in, smart_open.for_writing(compressed_path) as f_out:
-        shutil.copyfileobj(f_in, f_out)
-      print file_path, compressed_path
-      format = Registry.find(compressed_path)
-      try:
-        i = format(compressed_path)
-      except Exception:
-        print 'Error reading compressed file: %s' %compressed_path
-        import traceback
-        traceback.print_exc()
-      else:
-        print 'Successfully read compressed file: %s' %compressed_path
-        det = i.get_detector()
-        if det is not None:
-          size = det[0].get_image_size()
-        b = i.get_beam()
-        g = i.get_goniometer()
-        s = i.get_scan()
-        try:
-          d = i.get_raw_data()
-        except IOError:
-          pass
-
-def tst_dxtbx_models():
-
-  from dxtbx.tests.test_beam import test_beam
-  test_beam()
-
-  from dxtbx.tests.test_detector import test_detector
-  test_detector()
-
-  from dxtbx.tests.test_goniometer import test_goniometer
-  test_goniometer()
-
-  from dxtbx.tests.test_goniometer import test_multi_axis_goniometer
-  test_multi_axis_goniometer()
-
-  from dxtbx.tests.test_scan import test_scan
-  test_scan()
-
-def tst_sweep():
-
-  from dxtbx.sweep_filenames import template_regex
-
-  questions_answers = {
-      'foo_bar_001.img':'foo_bar_###.img',
-      'foo_bar001.img':'foo_bar###.img',
-      'foo_bar_1.8A_001.img':'foo_bar_1.8A_###.img',
-      'foo_bar.001':'foo_bar.###',
-      'foo_bar_001.img1000':'foo_bar_###.img1000',
-      'foo_bar_00001.img':'foo_bar_#####.img'
-      }
-
-  for filename in questions_answers:
-    answer = template_regex(filename)
-    assert answer[0] == questions_answers[filename]
-
-if __name__ == '__main__':
-  tst_dxtbx()
-  tst_dxtbx_models()  # these tests are failing in the misc_build.  Disable them until they can be debugged.
-  #tst_dxtbx_compressed()
